+++ mozdef_util/mozdef_util/elasticsearch_client.py
@@ -53,7 +53,11 @@ class ElasticsearchClient():
         self.es_connection.indices.delete(index=index_name, ignore=ignore_codes)
 
     def get_indices(self):
-        return list(self.es_connection.indices.stats()['indices'].keys())
+        # Includes open and closed indices
+        return list(self.es_connection.indices.get_alias('*', params=dict(expand_wildcards='all')).keys())
+
+    def get_open_indices(self):
+        return list(self.es_connection.indices.get_alias('*', params=dict(expand_wildcards='open')).keys())
 
     def index_exists(self, index_name):
         return self.es_connection.indices.exists(index_name)